requirements.txt
aggdraw==1.3.16
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.83
botocore==1.29.83
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comtypes==1.1.14
cryptography==39.0.2
docopt==0.6.2
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
imageio==2.26.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jwt==1.3.1
lazy_loader==0.1
MarkupSafe==2.1.2
multidict==6.0.4
networkx==3.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psd-tools==1.9.24
pycparser==2.21
python-dateutil==2.8.2
PyWavelets==1.4.1
s3transfer==0.6.0
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
tifffile==2023.2.28
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
